cli: extract language functions
